rename faqs service
